pyproject.toml
[project]
name = "FlipMarauder-WiFi"
version = "0.1.0"
description = "Passive Wi-Fi access point scanner built on top of Flipper Zero + ESP32 WiFi-Module V1 with Marauder Firmware. Collects, parses and persists real-time 802.11 beacon data through a USB-UART bridge."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "textual>=0.71.0",
    "pyserial>=3.5",
    "psutil>=7.2.2",
]
